status points cast; sync stats with player session obj on WoM usage to avoid data loss

+++ game-server/src/main/java/com/jftse/emulator/server/core/packets/player/S2CPlayerStatusPointChangePacket.java
@@ -56,6 +56,6 @@ public class S2CPlayerStatusPointChangePacket extends Packet {
             this.write((byte) 0);
         }
 
-        this.write(player.getStatusPoints());
+        this.write((byte) player.getStatusPoints());
     }
 }
